Makefile and weeklyCompile* to correctly phase .tex creation
This commit does a few things:

1. It adds a try/catch to weeklyCompile*.py so that if the .tex file doesn't
exist (like on the first `make`), it won't error but will allow the file to be
created. This was just a bug in the first version of this.

2. It makes explicit the phasing between `index` (a new name for compile),
which should be responsible for *all files generated from data* rather than a
one-to-one correspondence to a source file.

3. It adds a self-documenting first rule to guide users towards the README and
their likely intended action. (Note that the @ just makes it so `make` doesn't
print the echo command line in addition to it's output; that's just a special
symbol at the start of a normal unix `echo` command)

4. It cleans up a few make rule paths that were pointing at generated/output
instead of generated/output/app|topic, which was necessary for disambiguating
how to build things in output/ (and also if a topic and application end up with
the same name some day, it avoids that clash)

5. Updates the github build rule to use the combined index; website build.

(2) is important and warrants more explanation.

There was an issue with our makefile where we would have had to run `make`
*twice* in order to get the full build. The first make would expand e.g.
topicLatex and find no files in generated/notes/topic/*.tex, *before* running
any rules. This is because expansion is a sort of compile-time thing in make,
so the make tree is built by using the initial filesystem state. Actually
running the commands in rules is a runtime thing. The “code smell” for this is
any (wildcard generated/**) stuff, which refers to an intermediate file that
won't exist when the tree is created the first time.

I've done some poking around, and come to the conclusion that it's just
untenably difficult to get `make` to do this expansion “at runtime.” I've also
come to the conclusion that it's OK, we actually do just have two phases to our
build. The first I've renamed to `index` and the second is still `website`.
This isn't an unreasonable thing, and it's not the end of the world to have two
rules that collaborate to build as long as it's well-documented.

As a useful semantic summary, `make index; make website` is not the same thing
as a rule `all: index website` using `make all`, because the file system
inspected for `make website` in the first rule contains updates from `make
index`, and does not in the hypothetical `all` rule.