Make assignments act more like lessons than topic/app
The issue is that assignments have their source written directly by the course author, which is more like lessons/, but the assignment make rules were analogous to those of topic/app. This caused them to “miss” when creating targets from e.g. in (wildcard generated/notes/assignments/*.tex) because `make index` doesn't populate that.

The thing to be suspicious of is rules that use (wildcard generated/...) when we aren't totally sure that what comes after generated/... is populated by `make index`.